Started re-factoring code to generic landscape generators.
A generator is merely creating a noise map (with a client supplied noise function), runs a filter on it and then places things on it, depending on the current arguments.

The whole code is still very much subject to change, but I firmly believe that without all this redundancy it's going to be easier to add new, custom plant functions for example. Although it's not necessarily limited to plants at all.